Move oidc helpers to separate package
